% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/african_name_generator.R
\name{african_name}
\alias{african_name}
\title{Find out what your African name is!}
\usage{
african_name(name, boy = TRUE)
}
\arguments{
\item{name}{A character that is your current, non-South African name}

\item{boy}{A logical which is either TRUE if you are a boy, and FALSE if you are a girl or just not a boy}
}
\value{
There is no return statement, it outputs a string with the user's name
}
\description{
Find out what your African name is!
}
\examples{
african_name("ben")
african_name("stacy", boy = FALSE)
}
